:recycle: :boom: use interfaces for all PDF operations
