+++ src/main/java/com/mindee/image/ImageExtractor.java
@@ -4,7 +4,8 @@ import com.mindee.geometry.Bbox;
 import com.mindee.geometry.PositionDataField;
 import com.mindee.input.InputSourceUtils;
 import com.mindee.input.LocalInputSource;
-import com.mindee.pdf.PDFUtils;
+import com.mindee.pdf.PDFBoxApi;
+import com.mindee.pdf.PDFOperation;
 import com.mindee.pdf.PdfPageImage;
 import java.awt.image.BufferedImage;
 import java.io.ByteArrayInputStream;
@@ -21,29 +22,13 @@ public class ImageExtractor {
   private final String filename;
   private final String saveFormat;
 
-  /**
-   * Init from a path.
-   *
-   * @param filePath Path to the file.
-   * @throws IOException Throws if the file can't be accessed.
-   */
-  public ImageExtractor(String filePath) throws IOException {
-    this(new LocalInputSource(filePath));
-  }
-
-  /**
-   * Init from a {@link LocalInputSource}.
-   *
-   * @param source The local source.
-   * @throws IOException Throws if the file can't be accessed.
-   */
-  public ImageExtractor(LocalInputSource source) throws IOException {
+  public ImageExtractor(LocalInputSource source, PDFOperation pdfOperation) throws IOException {
     this.filename = source.getFilename();
     this.pageImages = new ArrayList<>();
 
     if (source.isPdf()) {
       this.saveFormat = "jpg";
-      var pdfPageImages = PDFUtils.pdfToImages(source);
+      var pdfPageImages = pdfOperation.pdfToImages(source);
       for (PdfPageImage pdfPageImage : pdfPageImages) {
         this.pageImages.add(pdfPageImage.getImage());
       }
@@ -54,6 +39,16 @@ public class ImageExtractor {
       var input = new ByteArrayInputStream(source.getFile());
       this.pageImages.add(ImageIO.read(input));
     }
+  }
+
+  /**
+   * Init from a {@link LocalInputSource}.
+   *
+   * @param source The local source.
+   * @throws IOException Throws if the file can't be accessed.
+   */
+  public ImageExtractor(LocalInputSource source) throws IOException {
+    this(source, new PDFBoxApi());
   }
 
   /**
