+++ src/test/java/com/mindee/pdf/PDFOperationTest.java
@@ -3,18 +3,73 @@ package com.mindee.pdf;
 import static com.mindee.TestingUtilities.getResourcePath;
 
 import com.mindee.MindeeException;
+import com.mindee.input.LocalInputSource;
 import com.mindee.input.PageOptions;
 import com.mindee.input.PageOptionsOperation;
+import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Random;
+import org.apache.pdfbox.pdmodel.PDDocument;
+import org.apache.pdfbox.pdmodel.PDPage;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 public class PDFOperationTest {
 
   private final PDFOperation pdfOperation = new PDFBoxApi();
+
+  @Test
+  public void shouldConvertSinglePageToJpg() throws IOException {
+    LocalInputSource source = new LocalInputSource(
+      "src/test/resources/file_types/pdf/multipage.pdf"
+    );
+    PdfPageImage pdfPageImage = pdfOperation.pdfPageToImage(source, 3);
+    Assertions.assertNotNull(pdfPageImage.getImage());
+    Assertions.assertEquals(pdfPageImage.asInputSource().getFilename(), pdfPageImage.getFilename());
+    pdfPageImage.writeToFile("src/test/resources/output/");
+    Assertions
+      .assertTrue(
+        Files.exists(Paths.get("src/test/resources/output/" + pdfPageImage.getFilename()))
+      );
+  }
+
+  @Test
+  public void shouldConvertAllPagesToJpg() throws IOException {
+    LocalInputSource source = new LocalInputSource(
+      "src/test/resources/file_types/pdf/multipage.pdf"
+    );
+    List<PdfPageImage> pdfPageImages = pdfOperation.pdfToImages(source);
+    for (PdfPageImage pdfPageImage : pdfPageImages) {
+      Assertions.assertNotNull(pdfPageImage.getImage());
+      Assertions
+        .assertEquals(pdfPageImage.asInputSource().getFilename(), pdfPageImage.getFilename());
+      pdfPageImage.writeToFile("src/test/resources/output/");
+      Assertions
+        .assertTrue(
+          Files.exists(Paths.get("src/test/resources/output/" + pdfPageImage.getFilename()))
+        );
+    }
+  }
+
+  @Test
+  public void givenADocument_whenPageCounted_thenReturnsCorrectPageCount() throws IOException {
+    PDDocument document = new PDDocument();
+    int random = new Random().nextInt(30);
+    for (int i = 0; i < random; i++) {
+      PDPage page = new PDPage();
+      document.addPage(page);
+    }
+    document.save("src/test/resources/output/test.pdf");
+    document.close();
+    File file = getResourcePath("output/test.pdf").toFile();
+    LocalInputSource source = new LocalInputSource(file);
+    Assertions.assertEquals(random, pdfOperation.getNumberOfPages(source));
+    file.delete();
+  }
 
   @Test
   public void givenADocumentAndPageToKeep_whenSplit_thenReturnsOnlyKeptPage() throws IOException {
@@ -25,8 +80,7 @@ public class PDFOperationTest {
       .build();
 
     byte[] fileBytes = Files.readAllBytes(getResourcePath("file_types/pdf/multipage.pdf"));
-    SplitQuery splitQuery = new SplitQuery(fileBytes, pageOptions);
-    SplitPDF splitPdf = pdfOperation.split(splitQuery);
+    SplitPDF splitPdf = pdfOperation.split(fileBytes, pageOptions);
 
     Assertions.assertNotNull(splitPdf);
     Assertions.assertNotNull(splitPdf.getFile());
@@ -45,11 +99,8 @@ public class PDFOperationTest {
       .operation(PageOptionsOperation.KEEP_ONLY)
       .build();
 
-    SplitQuery splitQuery = new SplitQuery(
-      Files.readAllBytes(getResourcePath("file_types/pdf/multipage.pdf")),
-      pageOptions
-    );
-    SplitPDF splitPdf = pdfOperation.split(splitQuery);
+    SplitPDF splitPdf = pdfOperation
+      .split(Files.readAllBytes(getResourcePath("file_types/pdf/multipage.pdf")), pageOptions);
 
     Assertions.assertNotNull(splitPdf);
     Assertions.assertNotNull(splitPdf.getFile());
@@ -64,11 +115,8 @@ public class PDFOperationTest {
       .operation(PageOptionsOperation.REMOVE)
       .build();
 
-    SplitQuery splitQuery = new SplitQuery(
-      Files.readAllBytes(getResourcePath("file_types/pdf/multipage.pdf")),
-      pageOptions
-    );
-    SplitPDF splitPdf = pdfOperation.split(splitQuery);
+    SplitPDF splitPdf = pdfOperation
+      .split(Files.readAllBytes(getResourcePath("file_types/pdf/multipage.pdf")), pageOptions);
 
     Assertions.assertNotNull(splitPdf);
     Assertions.assertNotNull(splitPdf.getFile());
@@ -83,12 +131,12 @@ public class PDFOperationTest {
       .operation(PageOptionsOperation.REMOVE)
       .build();
 
-    SplitQuery splitQuery = new SplitQuery(
-      Files.readAllBytes(getResourcePath("file_types/receipt.jpg")),
-      pageOptions
-    );
-
-    Assertions.assertThrows(MindeeException.class, () -> pdfOperation.split(splitQuery));
+    Assertions
+      .assertThrows(
+        MindeeException.class,
+        () -> pdfOperation
+          .split(Files.readAllBytes(getResourcePath("file_types/receipt.jpg")), pageOptions)
+      );
   }
 
   @Test
@@ -100,11 +148,11 @@ public class PDFOperationTest {
       .onMinPages(5)
       .build();
 
-    SplitQuery splitQuery = new SplitQuery(
-      Files.readAllBytes(getResourcePath("file_types/pdf/multipage_cut-2.pdf")),
-      pageOptions
-    );
-    SplitPDF splitPdf = pdfOperation.split(splitQuery);
+    SplitPDF splitPdf = pdfOperation
+      .split(
+        Files.readAllBytes(getResourcePath("file_types/pdf/multipage_cut-2.pdf")),
+        pageOptions
+      );
 
     Assertions.assertNotNull(splitPdf);
     Assertions.assertNotNull(splitPdf.getFile());
@@ -119,11 +167,8 @@ public class PDFOperationTest {
       .operation(PageOptionsOperation.KEEP_ONLY)
       .build();
 
-    SplitQuery splitQuery = new SplitQuery(
-      Files.readAllBytes(getResourcePath("file_types/pdf/multipage.pdf")),
-      pageOptions
-    );
-    SplitPDF splitPdf = pdfOperation.split(splitQuery);
+    SplitPDF splitPdf = pdfOperation
+      .split(Files.readAllBytes(getResourcePath("file_types/pdf/multipage.pdf")), pageOptions);
 
     Assertions.assertNotNull(splitPdf);
     Assertions.assertNotNull(splitPdf.getFile());
